Fix typo and discrepancies
`15100811449` is equivalent to `2017-11-07T18:59:04.9Z` but in binary form it is represented as `001110000100000101000001100010111001`, but not `001110000100000101000100000000110010`

VendorId=9 is encoded `00000000000001001` but according to the spec it should consist of 16 bits. Removing extra 0 in the head of the number.

Full consent string in binary format is `0000010011100001000001010001000000001100100011100001000001010001000000001100100000000001110000000000010000110001000011010000000010001110000000000000000000000000011111011011110000000000010000000000000100100000` (assembled from values mentioned in example table and padded to the nearest multiple 8 bits). Transforming to bytes and applying Base64 URL encoding gives the value of `BOEFEAyOEFEAyAHABDENAI4AAAB9vABAASA` not `BOJObISOJObISAABAAENAA4AAAAAoAAA` (validated via http://acdn.origin.appnexus.net/cmp/docs/#/tools/vendor-cookie-decoder)